requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.1.0.25
paho-mqtt==1.5.0
Pillow==7.0.0
protobuf==3.11.3
PyYAML==5.3
requests==2.20.0
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.24.3
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.12.0
